remove pong and use multiple listeners
